Reorganise: tests dir structure match src dir structure
